Moved BrewingRecipeCWrapper into own wrapper package (maybe other wrappers will follow sometime...)
