+++ CraftTweaker2-MC1120-Mod-JEI/src/main/java/crafttweaker/mods/jei/recipeWrappers/BrewingRecipeCWrapper.java
@@ -1,11 +1,17 @@
-package crafttweaker.mods.jei.actions;
+package crafttweaker.mods.jei.recipeWrappers;
 
 import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Collectors;
+
 import crafttweaker.mc1120.brewing.MultiBrewingRecipe;
+import crafttweaker.mods.jei.JEIAddonPlugin;
 import mezz.jei.api.ingredients.IIngredients;
+import mezz.jei.api.recipe.IRecipeWrapper;
+import mezz.jei.api.recipe.VanillaRecipeCategoryUid;
 import mezz.jei.plugins.vanilla.brewing.BrewingRecipeWrapper;
 import net.minecraft.item.ItemStack;
+import net.minecraftforge.common.brewing.BrewingRecipeRegistry;
 
 public class BrewingRecipeCWrapper extends BrewingRecipeWrapper  {
 	
@@ -28,9 +34,26 @@ public class BrewingRecipeCWrapper extends BrewingRecipeWrapper  {
 	}
 	
 	@Override
-	public List getInputs() {
+	public List<?> getInputs() {
 		// returns our inputs
 		return inputs;
 	}
+	
+	
+	
+	//Static Methods//	
+    public static List<IRecipeWrapper> createBrewingRecipes() {
+    	return BrewingRecipeRegistry.getRecipes()
+    					.stream()
+    					.filter(MultiBrewingRecipe.class::isInstance)
+    					.map(MultiBrewingRecipe.class::cast)
+    					.filter(MultiBrewingRecipe::isVisible)
+    					.map(BrewingRecipeCWrapper::new)
+    					.collect(Collectors.toList());
+    }
+    
+    public static void registerBrewingRecipe() {
+    	JEIAddonPlugin.modRegistry.addRecipes(createBrewingRecipes(), VanillaRecipeCategoryUid.BREWING);
+    }
 	
 }
